CleanUp and NoClobber Param

+++ Module/cChocoEx/Public/New-cChocoExSourceFile.ps1
@@ -4,12 +4,14 @@ function New-cChocoExSourceFile {
         # Path of Output File
         [Parameter(Mandatory)]
         [string]
-        $Path
+        $Path,
+        # NoClobber
+        [Parameter()]
+        [switch]
+        $NoClobber
     )
     
     begin {
-        #Gather Requested Values
-
         $ExportString = "@{`n"
         $Absent = New-Object System.Management.Automation.Host.ChoiceDescription '&Absent'
         $Present = New-Object System.Management.Automation.Host.ChoiceDescription '&Present'
@@ -77,6 +79,21 @@ function New-cChocoExSourceFile {
     }
     
     end {
-        $ExportString | Set-Content -Path $Path
+        try {
+            if ($NoClobber -and (Test-Path -Path $Path)) {
+                Write-Warning "File Already Exists and NoClobber Specified. Requesting Alternative Path"
+                $Path = Read-Host -Prompt "Path"
+                $ExportString | Set-Content -Path $Path
+            }
+            else {
+                $ExportString | Set-Content -Path $Path
+            }
+            $FullPath = (Get-Item -Path $Path).Fullname
+            Write-Host "File Wriiten to $FullPath"
+        }
+        catch {
+            $_.Exception.Message
+        }
+
     }
 }
